kernel: sucompat: use seccomp.mode for permission check
we zero it anyway on setuid, so we can repurpose it to speedup return checks.
we abuse the fact that task_struct is real hot.

upstream is doing the same thing anyway. set something on task_struct and use it to check.
ref: https://github.com/tiann/KernelSU/pull/2866

rebase of: https://github.com/backslashxx/KernelSU/commit/36348f2f50692893a98a30f140cb0898ba1462ea

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>