kernel: compat: uprev init_session_keyring pullout to < 5.2
well, it is no longer __rcu on 5.2+ so this is just my assumption that
this is really a < 5.2 issue. its probably just because of ACK backports
that it works without it on 4.14 and 4.19 MSM/ACK.

This is just a hypothesis as hisi and samsung kernels tend to have the issue.
Well, they dont follow ACK 1:1 anyway.

checking and adding a key is not that much of an issue, since this
is mostly used on kthreads.

This supercedes:
- kernel: expose allowlist workaround as Kconfig option
- https://github.com/backslashxx/KernelSU/commit/4c510ad8608cf2acffde87e3aaadca4dfd6a8b57

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>