kernel: sulog: basic ringbuffer su log
This is mostly to log su and escalation attempts from userspace.

this is a very basic ringbuffer implementation.

on this one we keep 5 bytes per entry and 100 entries max.
since this is pretty small, we can put this fully on stack.

userspace can deduce as we also send next index to it.
which makes it so that this is also the oldest entry.
next -1 is the latest entry.

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>